requirements.txt
numpy==1.19.5
toml
opencv-python
tensorrt==8.5.3.1
cuda-python==12.1.0
